Added search querries
