--- src/components/common/fetch/Fetch.css
@@ -1,23 +1,0 @@
-.card-image {
-    cursor: pointer;
-    background: rgb(231, 231, 231);
-    border-radius: 20px;
-    margin: 2rem 4rem;
-    overflow: hidden;
-    box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-    padding: 1rem 2rem;
-    transition: all 0.3s ease;
-    width:500px;
-    
-}
-
-.card-image img {
-    width:100%;
-    height:100%;
-    object-fit: cover;
-    border-radius:4px;
-}
-
-.card-image:hover {
-    transform: scale(1.1);
-}
