--- src/components/common/fetch/index.js
@@ -1,3 +1,0 @@
-import Fetch from "./Fetch";
-
-export default Fetch;
