+++ src/components/common/index.js
@@ -2,8 +2,8 @@ import Header from './header';
 import Navbar from './navbar';
 import Search from './search';
 import Logo from './logo';
-import Filter from './filter';
-import Fetch from './fetch';
+
+
 
 
 
@@ -14,8 +14,8 @@ export {
     Navbar,
     Search,
     Logo,
-    Filter,
-    Fetch
+    
+
