+++ src/components/common/navbar/index.js
@@ -1,6 +1,7 @@
 
 
 import Navbar from './Navbar';
+
 const hamburger = document.querySelector(".hamburger");
 const navMenu = document.querySelector(".nav-menu");
 
@@ -13,7 +14,5 @@ hamburger.addEventListener("click", () => {
     hamburger.classList.remove("active");
     navMenu.classList.remove("active");
   }))
-
-
 
 export default Navbar;
