+++ src/components/common/search/Search.css
@@ -6,7 +6,7 @@
 
 
 
-.form-wrapper input {
+.search-wrapper input {
     width:100%;
     max-width: 900px;
     border-radius: 60px;
@@ -15,5 +15,43 @@
     border:0;
     outline: none;
     padding: 15px 20px;
+    margin:2px;
+}
+.card-image {
+    cursor: pointer;
+    background: rgb(231, 231, 231);
+    border-radius: 20px;
+    margin: 2rem 4rem;
+    overflow: hidden;
+    box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
+    padding: 1rem 2rem;
+    transition: all 0.3s ease;
+    width:500px;
+    
+}
+
+.card-image img {
+    width:100%;
+    height:100%;
+    object-fit: cover;
+    border-radius:4px;
+}
+
+.card-image:hover {
+    transform: scale(1.1);
+}
+
+.visually-hidden {
+    border: 0 !important;
+    clip: rect(1px, 1px, 1px, 1px) !important;
+    -webkit-clip-path: inset(50%) !important;
+    clip-path: inset(50%) !important;
+    height: 1px !important;
+    margin: -1px !important;
+    overflow: hidden !important;
+    padding: 0 !important;
+    position: absolute !important;
+    width: 1px !important;
+    white-space: nowrap !important;
 }
