+++ src/index.css
@@ -84,7 +84,7 @@ picture, svg {
 }
 
 .content-search {
-  height: 700px;
+  height: 1800px;
   background-image: linear-gradient(rgba(0,8,51,0.9),rgba(0,8,51,0.9)) ,url('./virginia.jpg');
   background-repeat: no-repeat;
   background-size: cover;
